+++ gpaw/inducedfield/inducedfield_base.py
@@ -134,6 +134,7 @@ class BaseInducedField(object):
             self.na = len(self.atoms.get_atomic_numbers())        # !
             self.gd = self.density.gd                             # !
             self.stencil = self.density.stencil
+            self.log = paw.log
 
         if allocate:
             self.allocate()
@@ -291,7 +292,7 @@ class BaseInducedField(object):
         ng = r.ng
 
         # Background electric field
-        Fbgef_v = r.Fbgef_v
+        self.Fbgef_v = r.Fbgef_v
 
         if self.has_paw:
             # Test dimensions
@@ -301,13 +302,12 @@ class BaseInducedField(object):
                 raise IOError('nspins is incompatible with calculator')
             if (ng != self.gd.get_size_of_global_array()).any():
                 raise IOError('grid is incompatible with calculator')
-            if (Fbgef_v != self.Fbgef_v).any():
+            if (self.Fbgef_v != self.paw.kick_strength).any():
                 raise IOError('kick is incompatible with calculator')
         else:
             # Construct objects / assign values without paw
             self.na = na
             self.nspins = nspins
-            self.Fbgef_v = Fbgef_v
 
             from ase.io.trajectory import read_atoms
             self.atoms = read_atoms(r.atoms)