+++ gpaw/test/fd_ops/test_nabla.py
@@ -39,4 +39,4 @@ def test_fd_ops_nabla():
     d2 = TestSetup().get_derivative_integrals(rgd, g, np.zeros_like(g))
     if 0 in d1:
         print(abs(d1[0] - d2).max())
-        assert abs(d1[0] - d2).max() < 2e-6
+        assert abs(d1[0] - d2).max() < 2e-9