+++ doc/releasenotes.rst
@@ -12,6 +12,18 @@ Git master branch
 
 * Corresponding ASE release: ASE-3.21.1b1
 
+* The :ref:`time-propagation TDDFT (fd-mode) <timepropagation>` calculator
+  refactored and observer support generalized.
+
+  * The dipole moment output and restart file parameters are
+    deprecated; use the corresponding observers instead.
+    See the updated :ref:`documentation <timepropagation>`.
+
+  * The observers for :ref:`inducedfield` need now to be defined before
+    the kick instead of after it.
+
+  * Corresponding updates for :ref:`qsfdtd` and :ref:`hybridscheme`.
+
 * The documentation and tutorial for :ref:`lrtddft2` updated.
 
 * True occupation numbers are now printed in the text output for the
@@ -37,6 +49,24 @@ Git master branch
 
 * Added documentation for :ref:`elph` and added support for
   spin-polarized systems.
+
+* Implemented multiple orbital hubbard U corrections (EX: for correction
+  of both p and d orbitals on transition metals)
+
+* There used to be two versions of the GPAW web-page which was quite
+  confusing.  The https://wiki.fysik.dtu.dk/gpaw/dev/ web-page has now been
+  dropped.  There is now only https://wiki.fysik.dtu.dk/gpaw/ and it documents
+  the use of the in development version of GPAW.
+
+* You can now change all sorts of things about how the SCF cycle decides it
+  is converged. You can specify new, non-default convergence keywords like
+  ``work function`` or ``minimum iterations``, you can change how default
+  convergence keywords behave (like changing how many past energies the
+  ``energy`` criterion examines), and you can even write your own custom
+  convergence criteria. See :ref:`custom_convergence`.
+
+* The SCF output table has been simplified, and a letter "c" now appears
+  next to converged items.
 
 
 Version 21.1.0